refactor(LogoutButton): Migrate to TypeScript
Converted from JavaScript to TypeScript and simplified logout button component

Closes #206
