+++ components/auth/index.ts
